added login functionality pt2
