+++ src/backend/app/views.py
@@ -6,6 +6,7 @@ from rest_framework import status
 from app.models import User
 from rest_framework.permissions import AllowAny, IsAuthenticated
 from rest_framework_simplejwt.tokens import RefreshToken
+from django.contrib.auth import authenticate
 import requests
 
 def index(request):
@@ -55,19 +56,21 @@ class create_user(APIView):
             return Response({'error': f'Missing field: {e}'}, status=status.HTTP_400_BAD_REQUEST)
         except Exception as e:
             return Response({'error': str(e)}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
-class login_user(APIview):
+
+class login_user(APIView):
     permission_classes = [AllowAny]
 
     def post(self, request):
         username = request.data.get('username')
         password = request.data.get('password')
 
-        user = IsAuthenticated(username=username, password=password)
+        user = authenticate(username=username, password=password)
         if user: 
             refresh = RefreshToken.for_user(user)
-            return Response({'access': str(refresh.access_token),
-            'refresh': str(refresh)})
-        return Response({'error': 'Invalid credentials'}, status-status.HTTP_401_UNAUTHORIZED)
+            return Response({
+                'access': str(refresh.access_token),
+                'refresh': str(refresh)}, status=status.HTTP_200_OK)
+        return Response({'error': 'Invalid credentials'}, status=status.HTTP_401_UNAUTHORIZED)
 # Testing JWT requirement
 class protected_view(APIView):
     permission_classes = [IsAuthenticated]
