+++ src/backend/proj/urls.py
@@ -1,7 +1,7 @@
 from django.contrib import admin
 from django.urls import path, re_path, include
 from app import views
-from .views import login_user
+from app.views import login_user
 from rest_framework import routers, serializers, viewsets
 from rest_framework_simplejwt.views import (
     TokenObtainPairView,
@@ -15,6 +15,6 @@ urlpatterns = [
     path('api/token/refresh/', TokenRefreshView.as_view()),
     path('api/token/verify/', TokenVerifyView.as_view()),
     path('api/register/', views.create_user.as_view()),
-    path('login/', login_user.as_view()),
+    path('api/login/', login_user.as_view()),
     path('api/protected/', views.protected_view.as_view()),
 ]
